pyproject.toml
[project]
name = "computational-ml-lab"
version = "0.1.0"
description = "Physics/maths-driven ML engineering: inference, simulation, deep learning, and MLOps"
authors = [
    {name = "David Gisbert Ortiz", email = "davidgisbertortiz@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "lightgbm>=4.1.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "mlflow>=2.8.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.10.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]
extras = [
    "faiss-cpu>=1.7.4",
    "xgboost>=2.0.0",
    "plotly>=5.17.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["modules.*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["N802", "N803"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["modules"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=modules",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["modules"]
omit = ["*/tests/*", "*/notebooks/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
